created AndroidNotificationHelper for managing Notifications
