+++ app/src/main/java/com/raghav/spacedawnv2/di/MiscModule.kt
@@ -3,10 +3,13 @@ package com.raghav.spacedawnv2.di
 import android.content.Context
 import android.media.MediaPlayer
 import android.provider.Settings
+import com.raghav.spacedawnv2.domain.util.NotificationHelper
 import com.raghav.spacedawnv2.domain.util.ReminderScheduler
+import com.raghav.spacedawnv2.util.AndroidNotificationHelper
 import com.raghav.spacedawnv2.util.AndroidReminderScheduler
 import com.raghav.spacedawnv2.util.CoroutineDispatchers
 import com.raghav.spacedawnv2.util.DispatchersProvider
+import dagger.Binds
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -16,20 +19,25 @@ import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
-object MiscModule {
+abstract class MiscModule {
 
-    @Provides
-    @Singleton
-    fun provideReminderScheduler(@ApplicationContext context: Context): ReminderScheduler {
-        return AndroidReminderScheduler(context)
+    companion object {
+        @Provides
+        @Singleton
+        fun provideReminderScheduler(@ApplicationContext context: Context): ReminderScheduler {
+            return AndroidReminderScheduler(context)
+        }
+
+        @Singleton
+        @Provides
+        fun provideDispatchers(): DispatchersProvider = CoroutineDispatchers()
+
+        @Singleton
+        @Provides
+        fun provideMediaPlayer(@ApplicationContext context: Context): MediaPlayer =
+            MediaPlayer.create(context, Settings.System.DEFAULT_ALARM_ALERT_URI)
     }
 
-    @Singleton
-    @Provides
-    fun provideDispatchers(): DispatchersProvider = CoroutineDispatchers()
-
-    @Singleton
-    @Provides
-    fun provideMediaPlayer(@ApplicationContext context: Context): MediaPlayer =
-        MediaPlayer.create(context, Settings.System.DEFAULT_ALARM_ALERT_URI)
+    @Binds
+    abstract fun bindNotificationHelper(notificationHelper: AndroidNotificationHelper): NotificationHelper
 }
